Edição item matricula.

+++ Academia.Io.Tests/Manipuladores/MatricularManipuladorTests.cs
@@ -25,7 +25,7 @@ namespace Academia.Io.Tests.Manipuladores
                 PercentualDesconto = 10,
                 ValorDesconto = Convert.ToDecimal("0"),
                 ValorMatricula = 0,
-                ValorContrato = Convert.ToDouble("560,00"),
+                ValorContrato = Convert.ToDecimal("560,00"),
                 DiaVencimento = 1,
                 Ano = 2019,
                 DataIncialPagamento = DateTime.Now,
